Move config-related in config.go
