+++ gitmux.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"context"
-	_ "embed"
 	"flag"
 	"fmt"
 	"io"
@@ -30,21 +29,6 @@ Options:
   -timeout DUR    exits if still running after given duration (ex: 2s, 500ms).
   -V              prints gitmux version and exits.
 `
-
-// Config configures output formatting.
-type Config struct{ Tmux tmux.Config }
-
-// default config (decoded in init)
-var defaultCfg Config
-
-//go:embed .gitmux.yml
-var cfgBytes []byte
-
-func init() {
-	if err := yaml.Unmarshal(cfgBytes, &defaultCfg); err != nil {
-		panic(fmt.Sprintf("default config is invalid: %v", err))
-	}
-}
 
 func parseOptions() (ctx context.Context, cancel func(), dir string, dbg bool, cfg Config) {
 	var (
